+++ test/framework/src/main/java/org/opensearch/search/aggregations/AggregatorTestCase.java
@@ -92,6 +92,7 @@ import org.opensearch.index.mapper.CompletionFieldMapper;
 import org.opensearch.index.mapper.ContentPath;
 import org.opensearch.index.mapper.DateFieldMapper;
 import org.opensearch.index.mapper.FieldAliasMapper;
+import org.opensearch.index.mapper.FieldCorrelationMapper;
 import org.opensearch.index.mapper.FieldMapper;
 import org.opensearch.index.mapper.GeoPointFieldMapper;
 import org.opensearch.index.mapper.GeoShapeFieldMapper;
@@ -184,6 +185,7 @@ public abstract class AggregatorTestCase extends OpenSearchTestCase {
         denylist.add(ObjectMapper.NESTED_CONTENT_TYPE); // TODO support for nested
         denylist.add(CompletionFieldMapper.CONTENT_TYPE); // TODO support completion
         denylist.add(FieldAliasMapper.CONTENT_TYPE); // TODO support alias
+        denylist.add(FieldCorrelationMapper.CONTENT_TYPE); // TODO support correlation
         TYPE_TEST_DENYLIST = denylist;
     }